requirements.txt
transformers==4.30.0
tokenizer
accelerate
sentencepiece

overrides
pytorch-nlp
nltk
scipy
scikit-learn
tensorboard
pypinyin
statsmodels
lemminflect
num2words
word2number
language_tool_python
bert_score
editdistance
datasets
terminaltables
tensorflow_hub
tensorflow
deepspeed
openai
py-cpuinfo
lru-dict
flair
# torch._six = (bool,int)
fire
